- update to support x64 build - changed some vs project configs for parallel builds, dep flag, ... - use compiler warnning level4 - corrected cast to avoid 64bit pointer truncation - added some missing casts - provide empty string instead of false as return value - fixed warning level4 compiler warnings

+++ NppCommon/pipe.cpp
@@ -127,7 +127,7 @@ bool pipe::write(const void *aBuffer, int aBufSize)
 {
 	DWORD tWritten = 0;
 	const BOOL tResult = WriteFile(mPipe, aBuffer, aBufSize, &tWritten, NULL);
-	return TRUE == tResult && tWritten == aBufSize;
+	return TRUE == tResult && tWritten == static_cast<DWORD>(aBufSize);
 }
 
 void pipe::close()
